Reorganize interfaces in sub namespaces
