+++ src/Version10/Configuration/ConfigurationOptionInterface.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Phpcq\PluginApi\Version10;
+namespace Phpcq\PluginApi\Version10\Configuration;
+
+use Phpcq\PluginApi\Version10\Exception\InvalidConfigException;
 
 interface ConfigurationOptionInterface
 {
