+++ src/Version10/ConfigurationPluginInterface.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Phpcq\PluginApi\Version10;
 
+use Phpcq\PluginApi\Version10\Configuration\ConfigurationOptionsBuilderInterface;
 use Phpcq\PluginApi\Version10\Task\TaskInterface;
 
 interface ConfigurationPluginInterface extends PluginInterface
@@ -19,11 +20,11 @@ interface ConfigurationPluginInterface extends PluginInterface
      * Process plugin configuration and create task runners.
      *
      * @param array                $config      The plugin configuration.
-     * @param BuildConfigInterface $buildConfig The build configuration.
+     * @param EnvironmentInterface $buildConfig The build configuration.
      *
      * @return TaskInterface[]
      *
      * @psalm-return \Generator<int, TaskInterface>
      */
-    public function processConfig(array $config, BuildConfigInterface $buildConfig): iterable;
+    public function processConfig(array $config, EnvironmentInterface $buildConfig): iterable;
 }
