+++ src/Version10/Exception/InvalidConfigException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phpcq\PluginApi\Version10;
+namespace Phpcq\PluginApi\Version10\Exception;
 
 use RuntimeException;
