+++ src/Version10/Output/OutputTransformerFactoryInterface.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Phpcq\PluginApi\Version10;
+namespace Phpcq\PluginApi\Version10\Output;
+
+use Phpcq\PluginApi\Version10\Report\ToolReportInterface;
 
 interface OutputTransformerFactoryInterface
 {
