+++ src/Version10/Output/OutputTransformerInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phpcq\PluginApi\Version10;
+namespace Phpcq\PluginApi\Version10\Output;
 
 /**
  * This interface handles output transformations of tasks.
