+++ src/Version10/Report/ReportInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phpcq\PluginApi\Version10;
+namespace Phpcq\PluginApi\Version10\Report;
 
 interface ReportInterface
 {
