--- src/Version10/RuntimeException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Phpcq\PluginApi\Version10;
-
-class RuntimeException extends \RuntimeException implements Exception
-{
-}
