+++ src/Version10/Task/OutputWritingTaskInterface.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Phpcq\PluginApi\Version10\Task;
 
-use Phpcq\PluginApi\Version10\OutputInterface;
-use Phpcq\PluginApi\Version10\RuntimeException;
+use Phpcq\PluginApi\Version10\Output\OutputInterface;
+use Phpcq\PluginApi\Version10\Exception\RuntimeException;
 
 interface OutputWritingTaskInterface extends TaskInterface
 {
