+++ src/Version10/Task/ReportWritingTaskInterface.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Phpcq\PluginApi\Version10\Task;
 
-use Phpcq\PluginApi\Version10\RuntimeException;
-use Phpcq\PluginApi\Version10\ToolReportInterface;
+use Phpcq\PluginApi\Version10\Exception\RuntimeException;
+use Phpcq\PluginApi\Version10\Report\ToolReportInterface;
 
 interface ReportWritingTaskInterface extends TaskInterface
 {
