+++ tests/Version10/InvalidConfigExceptionTest.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Phpcq\PluginApi\Test\Version10;
 
-use Phpcq\PluginApi\Version10\Exception;
-use Phpcq\PluginApi\Version10\InvalidConfigException;
+use Phpcq\PluginApi\Version10\Exception\Exception;
+use Phpcq\PluginApi\Version10\Exception\InvalidConfigException;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \Phpcq\PluginApi\Version10\InvalidConfigException
+ * @covers \Phpcq\PluginApi\Version10\Exception\InvalidConfigException
  */
 final class InvalidConfigExceptionTest extends TestCase
 {
