+++ tests/Version10/ReportClosedExceptionTest.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Phpcq\PluginApi\Test\Version10;
 
-use Phpcq\PluginApi\Version10\Exception;
-use Phpcq\PluginApi\Version10\ReportClosedException;
+use Phpcq\PluginApi\Version10\Exception\Exception;
+use Phpcq\PluginApi\Version10\Exception\ReportClosedException;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \Phpcq\PluginApi\Version10\ReportClosedException
+ * @covers \Phpcq\PluginApi\Version10\Exception\ReportClosedException
  */
 final class ReportClosedExceptionTest extends TestCase
 {
