+++ tests/Version10/RuntimeExceptionTest.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Phpcq\PluginApi\Test\Version10;
 
-use Phpcq\PluginApi\Version10\Exception;
-use Phpcq\PluginApi\Version10\RuntimeException;
+use Phpcq\PluginApi\Version10\Exception\Exception;
+use Phpcq\PluginApi\Version10\Exception\RuntimeException;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \Phpcq\PluginApi\Version10\RuntimeException
+ * @covers \Phpcq\PluginApi\Version10\Exception\RuntimeException
  */
 final class RuntimeExceptionTest extends TestCase
 {
